pyproject.toml
[tool.poetry]
name = "anon-board"
version = "0.1.0"
description = ""
authors = ["Piti Ongmongkolkul <piti118@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
flask = "^3.0.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
